Delete translations, separate entity into balboa_entity.py
